pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.25.2",
    "openai>=1.82.0",
    "psycopg2-binary>=2.9.10",
    "python-telegram-bot==20.7",
]
